convert uuidcreator to ts
