+++ src/Jsoninput.svelte
@@ -4,7 +4,7 @@
   import { Container, Row, Col, Button, Form, FormGroup, FormText, Input, Label, Jumbotron, ListGroup, ListGroupItem } from "sveltestrap";
   
   import Uuidfinder, {findUuids} from './utils/Uuidfinder.svelte';
-  import Uuidcreator, {createUuid} from './utils/Uuidcreator.svelte';
+  import {createUuid} from './utils/uuidcreator';
   import {sampleJSON1, sampleJSON2, sampleJSON3} from './utils/samplejson.js';
   
   let textInput1;
